Renaming extraction to transformation and changing structure
