+++ core/transformation/towns_fund/round_1.py
@@ -14,7 +14,7 @@ from core.const import (
     LikelihoodEnum,
 )
 from core.controllers.mappings import INGEST_MAPPINGS
-from core.extraction.utils import drop_unnecessary_fhsf_data, extract_postcodes
+from core.transformation.utils import drop_unnecessary_fhsf_data, extract_postcodes
 
 
 def ingest_round_one_data_towns_fund(round_1_data: dict[str, pd.DataFrame]) -> dict[str, pd.DataFrame]:
