+++ tests/extraction_tests/conftest.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import pandas as pd
 import pytest
 
-from core.extraction import towns_fund_round_three as tf
+from core.transformation.towns_fund import round_3 as tf
 
 round_3_resources_mocks = Path(__file__).parent / "resources" / "mock_sheet_data" / "round_three"
