+++ tests/extraction_tests/test_round_three_extraction.py
@@ -12,8 +12,8 @@ from pandas.testing import assert_frame_equal
 from core.const import OUTCOME_CATEGORIES, OUTPUT_CATEGORIES
 from core.controllers.mappings import INGEST_MAPPINGS
 from core.exceptions import ValidationError
-from core.extraction import towns_fund_round_three as tf
-from core.extraction.towns_fund_round_two import ingest_round_two_data_towns_fund
+from core.transformation.towns_fund import round_3 as tf
+from core.transformation.towns_fund.round_2 import ingest_round_two_data_towns_fund
 
 resources = Path(__file__).parent / "resources"
 resources_mocks = resources / "mock_sheet_data" / "round_three"
@@ -65,7 +65,7 @@ def mock_ingest_full_sheet(
 
 
 @pytest.fixture
-@patch("core.extraction.towns_fund_round_three.TF_PLACE_NAMES_TO_ORGANISATIONS", {"Fake Town": "Fake Canonical Org"})
+@patch("core.transformation.towns_fund.round_3.TF_PLACE_NAMES_TO_ORGANISATIONS", {"Fake Town": "Fake Canonical Org"})
 def mock_ingest_full_extract(mock_ingest_full_sheet):
     """Setup mock of full spreadsheet extract."""
 
@@ -105,7 +105,7 @@ def test_project_lookup(mock_place_extract):
     }
 
 
-@patch("core.extraction.towns_fund_round_three.TF_PLACE_NAMES_TO_ORGANISATIONS", {"Fake Town": "Fake Canonical Org"})
+@patch("core.transformation.towns_fund.round_3.TF_PLACE_NAMES_TO_ORGANISATIONS", {"Fake Town": "Fake Canonical Org"})
 def test_extract_programme(mock_place_extract, mock_programme_lookup):
     """Test programme info extracted as expected."""
     test_extracted_programme_df = tf.extract_programme(mock_place_extract, mock_programme_lookup)
@@ -113,7 +113,7 @@ def test_extract_programme(mock_place_extract, mock_programme_lookup):
     assert_frame_equal(test_extracted_programme_df, expected_programme_df)
 
 
-@patch("core.extraction.towns_fund_round_three.TF_PLACE_NAMES_TO_ORGANISATIONS", {"Fake Town": "Fake Canonical Org"})
+@patch("core.transformation.towns_fund.round_3.TF_PLACE_NAMES_TO_ORGANISATIONS", {"Fake Town": "Fake Canonical Org"})
 def test_extract_organisation(mock_place_extract):
     """Test organisations details extracted as expected."""
     test_extracted_organisation_df = tf.extract_organisation(mock_place_extract)
